Incorporates the rest of the project content into the layout of the PortfolioPage template
